Refactor project structure and add  spinner
