+++ src/components/View/View.js
@@ -4,7 +4,6 @@ import { SIZE_RATIO } from '../../constants';
 
 const Container = glamorous.div(
   {
-    position: 'relative',
     margin: '2.5px',
     cursor: 'pointer',
     ':hover': {
@@ -14,6 +13,7 @@ const Container = glamorous.div(
   ({ height, width }) => {
     if (height && width) {
       const size = width * SIZE_RATIO / height;
+
       return {
         flexGrow: `${size}`,
         width: `${size}px`,
@@ -23,20 +23,10 @@ const Container = glamorous.div(
   },
 );
 
-const Divider = glamorous.i(
-  {
-    display: 'block',
-  },
-  ({ height, width }) => {
-    if (height && width) {
-      const size = height / width * 100;
-      return {
-        paddingBottom: `${size}%`,
-      };
-    }
-    return {};
-  },
-);
+const Image = glamorous.img({
+  width: '100%',
+  verticalAlign: 'bottom',
+});
 
 export default class View extends Component {
   constructor(props) {
@@ -55,6 +45,7 @@ export default class View extends Component {
       width: this.img.naturalWidth || this.img.width,
       height: this.img.naturalHeight || this.img.height,
     });
+    this.props.onChangeLoading();
   }
 
   setRef(img) {
@@ -66,18 +57,11 @@ export default class View extends Component {
     const { src, title, onClick } = this.props;
     return (
       <Container onClick={onClick} width={width} height={height}>
-        <Divider width={width} height={height} />
-        <img
+        <Image
           src={src}
           alt={title}
-          ref={this.setRef}
+          innerRef={this.setRef}
           onLoad={this.onLoad}
-          style={{
-            position: 'absolute',
-            top: '0',
-            width: '100%',
-            verticalAlign: 'bottom',
-          }}
         />
       </Container>
     );
